Change migrate file

+++ database/migrations/2024_08_25_125747_drop_product_id_column_from_size.php
@@ -1,28 +1,28 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
 
-return new class extends Migration
+class DropProductIdColumnFromSize extends Migration
 {
-    /**
-     * Run the migrations.
-     */
-    public function up(): void
+    public function up()
     {
-        Schema::table('sizes', function (Blueprint $table) {
-            $table->dropColumn('product_id');
-        });
+        if (Schema::hasColumn('sizes', 'product_id')) {
+            Schema::table('sizes', function (Blueprint $table) {
+                $table->dropForeign(['product_id']);
+                $table->dropColumn('product_id');
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
-    public function down(): void
+    public function down()
     {
-        Schema::table('sizes', function (Blueprint $table) {
-            $table->unsignedBigInteger('product_id');
-        });
+        if (! Schema::hasColumn('sizes', 'product_id')) {
+            Schema::table('sizes', function (Blueprint $table) {
+                $table->unsignedBigInteger('product_id')->after('id');
+                $table->foreign('product_id')->references('id')->on('products');
+            });
+        }
     }
-};
+}
